Added wall pos.txt and an array in level to add to
Created a text file called "wall_positions.txt" and wrote the positions of the walls in it, one per line.  Used the FileReader and BufferedReader classes to read the file and store the positions in an ArrayList. I replaced the conditional statement in the for loop with a call to the method of the ArrayList